requirements.txt
alembic==1.0.0
amqp==2.3.2
APScheduler==3.5.3
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.0
crypto==1.4.1
cryptography==2.3.1
cycler==0.10.0
Django==2.1.2
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-APScheduler==1.10.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
icalendar==4.0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
kombu==4.2.1
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.0
monotonic==1.5
Naked==0.1.31
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.2.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
schedule==0.5.0
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.12
tzlocal==1.5.1
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
